app/requirements.txt
aiohttp==3.8.5
pillow==10.4.0
sanic==23.6.0 
